changelog 0.3.3
--MAJOR--
updated enemyAI
-now enemy moves, has health, should attack, and the healthdot is updated, once the animations are in then it should also be animated.
added the objects necessary for the script to work.

+++ Reign_Piste/Assets/ASSETS/SCRIPTS/MISC/Health.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class Health : MonoBehaviour
 {
-    public int maxHealth = 1;
-    public int currentHealth;
+    public float maxHealth = 1f;
+    public float currentHealth;
     //fightcontroller will be made sometime soon(before wednesday for sure)
     private FightController fightController;
     
@@ -18,7 +18,7 @@ public class Health : MonoBehaviour
     }
     //yoinked the update method, dont need rn
     //gonna add one for taking damage though
-    public void TakeDamage(int damageAmount)
+    public void TakeDamage(float damageAmount)
     {
         currentHealth -= damageAmount;
         if (currentHealth <= 0)
